requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LLM and AI
groq==0.4.1
sentence-transformers==2.2.2

# Vector database and search
chromadb==0.4.18
faiss-cpu==1.7.4

# Document processing
pypdf2==3.0.1
markdown==3.5.1
beautifulsoup4==4.12.2
python-multipart==0.0.6

# Web scraping and HTTP
requests==2.31.0
aiofiles==23.2.0

# Data processing
pandas==2.1.4
numpy==1.24.3
python-docx==1.1.0

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAPI/Swagger parsing
openapi3-parser==1.1.14
pyyaml==6.0.1

# HTML parsing and processing
lxml==4.9.3

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Frontend
jinja2==3.1.2
